chore(react): improve
