+++ docs/content/docs/1.getting-started/2.installation/3.react.md
@@ -41,8 +41,8 @@ Import icons as React-components.
 ::component-example
 ---
 name: 'icon-import-example-react'
-collapse: true
-language: 'tsx'
+preview: false
+lang: 'tsx'
 ---
 ::
 
@@ -59,9 +59,9 @@ are used for load optimization, loading only the necessary parts of the code whe
 
 ::component-example
 ---
-name: 'icon-component-example-react'
-collapse: true
-language: 'tsx'
+name: 'icon-component-react-example'
+preview: false
+lang: 'tsx'
 ---
-[]::
+::
