+++ docs/package.json
@@ -14,7 +14,6 @@
     "@nuxt/content": "^3.8.0",
     "@nuxt/kit": "^4.2.1",
     "@bitrix24/b24icons-vue": "workspace:*",
-    "@bitrix24/b24icons-react": "workspace:*",
     "@bitrix24/b24ui-nuxt": "^2.1.1",
     "@nuxtjs/plausible": "^2.0.1",
     "@octokit/rest": "^22.0.1",
@@ -26,8 +25,6 @@
     "motion-v": "^1.7.4",
     "nuxt": "^4.2.1",
     "prettier": "^3.6.2",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
     "shiki-stream": "^0.1.3",
     "shiki-transformer-color-highlight": "^1.0.0",
     "sortablejs": "^1.15.6",
