+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@bitrix24/b24icons",
-  "version": "2.0.7",
+  "version": "2.0.8",
   "packageManager": "pnpm@10.21.0",
   "author": "Bitrix",
   "license": "MIT",
@@ -62,7 +62,7 @@
     "@nuxt/eslint-config": "^1.10.0",
     "@babel/core": "^7.28.5",
     "@babel/plugin-transform-react-jsx": "^7.27.1",
-    "@svgr/core": "^5.5.0",
+    "@svgr/core": "^8.1.0",
     "@vue/compiler-dom": "^3.5.24",
     "camelcase": "^9.0.0",
     "cross-env": "^10.1.0",
