+++ packages/@bitrix24-icons-nuxt/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@bitrix24/b24icons-nuxt",
-  "version": "2.0.7",
+  "version": "2.0.8",
   "author": "Bitrix",
   "license": "MIT",
   "description": "Bitrix24 SVG icons for development web-applications for Nuxt",
