+++ pnpm-workspace.yaml
@@ -1,8 +1,8 @@
 packages:
   - docs
   - packages/@bitrix24-icons-vue
-  - packages/@bitrix24-icons-react
   - packages/@bitrix24-icons-nuxt
+  - packages/@bitrix24-icons-react
 
 ignoredBuiltDependencies:
   - '@parcel/watcher'
